View内に定義していたコードをviewModelへ移行
